ecs parameters: move common mathods to a header only file
