+++ DataFormats/Parameters/src/GRPECSObject.cxx
@@ -108,47 +108,3 @@ GRPECSObject* GRPECSObject::loadFrom(const std::string& grpecsFileName)
   }
   return grp;
 }
-
-//_______________________________________________
-GRPECSObject::RunType GRPECSObject::string2RunType(const std::string& rts)
-{
-  int rt = -1;
-  for (int i = 0; i < int(RunType::NRUNTYPES); i++) {
-    if (rts == RunTypeNames[i]) {
-      rt = i;
-      break;
-    }
-  }
-  return GRPECSObject::RunType(rt);
-}
-
-//_______________________________________________
-std::string GRPECSObject::getRawDataPersistencyMode(const std::string& runType, bool imposeRaw)
-{
-  if (imposeRaw) {
-    return "raw";
-  }
-  std::string ret = "other";
-  auto rt = string2RunType(runType);
-  switch (rt) {
-    case RunType::PHYSICS:
-    case RunType::COSMICS:
-      ret = "raw";
-      break;
-    case RunType::PEDESTAL:
-    case RunType::PULSER:
-    case RunType::LASER:
-    case RunType::CALIBRATION_ITHR_TUNING:
-    case RunType::CALIBRATION_VCASN_TUNING:
-    case RunType::CALIBRATION_THR_SCAN:
-    case RunType::CALIBRATION_DIGITAL_SCAN:
-    case RunType::CALIBRATION_ANALOG_SCAN:
-    case RunType::CALIBRATION_FHR:
-    case RunType::CALIBRATION_ALPIDE_SCAN:
-    case RunType::CALIBRATION:
-      ret = "calib";
-    default:
-      break;
-  }
-  return ret;
-}
